+++ src/routes/clussster/index.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
 	import {base} from '$app/paths';
 
-	import SssmoothSnake from '$lib/sports/sssmooth/SssmoothSnake.svelte';
+	import ClusssterSnake from '$lib/sports/clussster/ClusssterSnake.svelte';
 	import TitleImage from '$lib/TitleImage.svelte';
 	import Footer from '$lib/Footer.svelte';
 
 	console.log(`base`, base);
 </script>
 
-<TitleImage alt="logo for svelte snake sports: sssmooth" src="{base}/assets/sssmooth.webp" />
+<TitleImage alt="logo for svelte snake sports: clussster" src="{base}/assets/clussster.webp" />
 
-<SssmoothSnake />
+<ClusssterSnake />
 
 <Footer />
